youbora integration cont.

+++ playkit/src/main/java/com/kaltura/playkit/plugins/Youbora/YouboraPlugin.java
@@ -126,7 +126,7 @@ public class YouboraPlugin extends PKPlugin {
     PKEvent.Listener eventListener = new PKEvent.Listener() {
         @Override
         public void onEvent(PKEvent event) {
-            Map<String, Object> opt  = YouboraConfig.setYouboraMediaDuration(pluginConfig, Long.valueOf(player.getDuration() / 1000).intValue());
+            Map<String, Object> opt  = YouboraConfig.setYouboraMediaDuration(pluginConfig, Long.valueOf(player.getDuration() / 1000).doubleValue());
             // Set options
             pluginManager.setOptions(opt);
         }
